Fix AllocationItemEditable comment
