+++ src/types.ts
@@ -1,0 +1,9 @@
+export type allocationType = {
+  name: string | undefined;
+  id: number | undefined;
+}
+
+export type editType = {
+  status: boolean;
+  allocationId: number | undefined;
+}
